requirements.txt
requests
flask
python-dotenv
tuya_connector
